+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/GoldService.scala
@@ -1458,13 +1458,13 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
   }
 
   class FinagledService(
-      iface: MethodPerEndpoint,
+      iface: MethodPerEndpoint with ServerValidationMixin,
       serverParam: RichServerParam)
     extends GoldService$FinagleService(iface, serverParam) {
 
     @deprecated("Use com.twitter.finagle.thrift.RichServerParam", "2017-08-16")
     def this(
-      iface: MethodPerEndpoint,
+      iface: MethodPerEndpoint with ServerValidationMixin,
       protocolFactory: org.apache.thrift.protocol.TProtocolFactory,
       serviceName: String = "GoldService"
     ) = this(iface, RichServerParam(protocolFactory, serviceName))