+++ scrooge-generator-tests/src/test/scala/com/twitter/scrooge/frontend/TypeResolverSpec.scala
@@ -386,6 +386,7 @@ class TypeResolverSpec extends Spec {
                   _,
                   _,
                   _,
+                  _,
                   _
                 ) => // pass
             case _ => fail()
@@ -396,6 +397,7 @@ class TypeResolverSpec extends Spec {
                   _,
                   _,
                   SetType(StructType(_, Some(SimpleID("typedef1", _))), None),
+                  _,
                   _,
                   _,
                   _,